refactor oauth token caching and authorization
The new oauth token caching code follows the suggestion in
https://github.com/golang/oauth2/issues/84#issuecomment-175834679, which
allows it to update the refresh token in addition to the access token.
This is required with Fitbit, but not with other providers, e.g. Google.

Authorization now uses an HTTP server on localhost and sets RedirectURL
to point to that server. That way, codes don’t need to be copy&pasted.
This isn’t implemented entirely by choice, AFAIR Fitbit started
requiring a valid RedirectURL.